pyproject.toml
[project]
name = "recurrentgemma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "accelerate>=1.12.0",
  "numpy<2.0.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "torch==2.10.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
  "torch<2.3.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "transformers>=5.2.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
  "transformers>=4.40.0,<5.0.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "easy-logging-formatter>=0.1.1",
  "datasets>=4.6.1",
]
